requirements.txt
llama-cpp-python
streamlit
python-dotenv
pydantic
